branch: feature/extract-node-logic - Move core functions to abstract class
